simplifying mapping json
